requirements.txt
asgiref==3.4.1
Django==4.0.4
django-cors-headers==3.10.0
django-filter==22.1
django-jazzmin==2.5.0
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
Pillow==8.4.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2022.1
